Remove unused LongDictionaryStreamReader
