--- presto-orc/src/main/java/io/prestosql/orc/reader/LongDirectStreamReader.java
@@ -1,192 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.prestosql.orc.reader;
-
-import io.prestosql.memory.context.LocalMemoryContext;
-import io.prestosql.orc.OrcCorruptionException;
-import io.prestosql.orc.StreamDescriptor;
-import io.prestosql.orc.metadata.ColumnEncoding;
-import io.prestosql.orc.stream.BooleanInputStream;
-import io.prestosql.orc.stream.InputStreamSource;
-import io.prestosql.orc.stream.InputStreamSources;
-import io.prestosql.orc.stream.LongInputStream;
-import io.prestosql.spi.block.Block;
-import io.prestosql.spi.block.BlockBuilder;
-import io.prestosql.spi.block.RunLengthEncodedBlock;
-import io.prestosql.spi.type.Type;
-import org.openjdk.jol.info.ClassLayout;
-
-import javax.annotation.Nullable;
-
-import java.io.IOException;
-import java.time.ZoneId;
-import java.util.List;
-
-import static com.google.common.base.MoreObjects.toStringHelper;
-import static io.airlift.slice.SizeOf.sizeOf;
-import static io.prestosql.orc.metadata.Stream.StreamKind.DATA;
-import static io.prestosql.orc.metadata.Stream.StreamKind.PRESENT;
-import static io.prestosql.orc.stream.MissingInputStreamSource.missingStreamSource;
-import static java.util.Objects.requireNonNull;
-
-public class LongDirectStreamReader
-        implements StreamReader
-{
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(LongDirectStreamReader.class).instanceSize();
-
-    private final StreamDescriptor streamDescriptor;
-
-    private int readOffset;
-    private int nextBatchSize;
-
-    private InputStreamSource<BooleanInputStream> presentStreamSource = missingStreamSource(BooleanInputStream.class);
-    @Nullable
-    private BooleanInputStream presentStream;
-    private boolean[] nullVector = new boolean[0];
-
-    private InputStreamSource<LongInputStream> dataStreamSource = missingStreamSource(LongInputStream.class);
-    @Nullable
-    private LongInputStream dataStream;
-
-    private boolean rowGroupOpen;
-
-    private final LocalMemoryContext systemMemoryContext;
-
-    public LongDirectStreamReader(StreamDescriptor streamDescriptor, LocalMemoryContext systemMemoryContext)
-    {
-        this.streamDescriptor = requireNonNull(streamDescriptor, "stream is null");
-        this.systemMemoryContext = requireNonNull(systemMemoryContext, "systemMemoryContext is null");
-    }
-
-    @Override
-    public void prepareNextRead(int batchSize)
-    {
-        readOffset += nextBatchSize;
-        nextBatchSize = batchSize;
-    }
-
-    @Override
-    public Block readBlock(Type type)
-            throws IOException
-    {
-        if (!rowGroupOpen) {
-            openRowGroup();
-        }
-
-        if (readOffset > 0) {
-            if (presentStream != null) {
-                // skip ahead the present bit reader, but count the set bits
-                // and use this as the skip size for the data reader
-                readOffset = presentStream.countBitsSet(readOffset);
-            }
-            if (readOffset > 0) {
-                if (dataStream == null) {
-                    throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
-                }
-                dataStream.skip(readOffset);
-            }
-        }
-
-        if (dataStream == null && presentStream != null) {
-            presentStream.skip(nextBatchSize);
-            Block nullValueBlock = RunLengthEncodedBlock.create(type, null, nextBatchSize);
-            readOffset = 0;
-            nextBatchSize = 0;
-            return nullValueBlock;
-        }
-
-        BlockBuilder builder = type.createBlockBuilder(null, nextBatchSize);
-        if (presentStream == null) {
-            if (dataStream == null) {
-                throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
-            }
-            dataStream.nextLongVector(type, nextBatchSize, builder);
-        }
-        else {
-            for (int i = 0; i < nextBatchSize; i++) {
-                if (presentStream.nextBit()) {
-                    type.writeLong(builder, dataStream.next());
-                }
-                else {
-                    builder.appendNull();
-                }
-            }
-        }
-
-        readOffset = 0;
-        nextBatchSize = 0;
-
-        return builder.build();
-    }
-
-    private void openRowGroup()
-            throws IOException
-    {
-        presentStream = presentStreamSource.openStream();
-        dataStream = dataStreamSource.openStream();
-
-        rowGroupOpen = true;
-    }
-
-    @Override
-    public void startStripe(ZoneId timeZone, InputStreamSources dictionaryStreamSources, List<ColumnEncoding> encoding)
-    {
-        presentStreamSource = missingStreamSource(BooleanInputStream.class);
-        dataStreamSource = missingStreamSource(LongInputStream.class);
-
-        readOffset = 0;
-        nextBatchSize = 0;
-
-        presentStream = null;
-        dataStream = null;
-
-        rowGroupOpen = false;
-    }
-
-    @Override
-    public void startRowGroup(InputStreamSources dataStreamSources)
-    {
-        presentStreamSource = dataStreamSources.getInputStreamSource(streamDescriptor, PRESENT, BooleanInputStream.class);
-        dataStreamSource = dataStreamSources.getInputStreamSource(streamDescriptor, DATA, LongInputStream.class);
-
-        readOffset = 0;
-        nextBatchSize = 0;
-
-        presentStream = null;
-        dataStream = null;
-
-        rowGroupOpen = false;
-    }
-
-    @Override
-    public String toString()
-    {
-        return toStringHelper(this)
-                .addValue(streamDescriptor)
-                .toString();
-    }
-
-    @Override
-    public void close()
-    {
-        systemMemoryContext.close();
-        nullVector = null;
-    }
-
-    @Override
-    public long getRetainedSizeInBytes()
-    {
-        return INSTANCE_SIZE + sizeOf(nullVector);
-    }
-}
